Added floating point support to colors to enhance dithering effect

+++ Project1/ColorQuantizer/Models/Color.cs
@@ -8,15 +8,15 @@ namespace ImageProcessing.Models
         /// <summary>
         /// First channel of the color, usually Red
         /// </summary>
-        public int Channel1 { get; set; }
+        public float Channel1 { get; set; }
         /// <summary>
         /// Second channel of the color, usually Green
         /// </summary>
-        public int Channel2 { get; set; }
+        public float Channel2 { get; set; }
         /// <summary>
         /// Third channel of the color, usually Blue
         /// </summary>
-        public int Channel3 { get; set; }
+        public float Channel3 { get; set; }
 
         public Color() { }
         /// <summary>
@@ -25,7 +25,7 @@ namespace ImageProcessing.Models
         /// <param name="channel1">first channel, usually Red, between -255 and 255</param>
         /// <param name="channel2">second channel, usually Green, between -255 and 255</param>
         /// <param name="channel3">third channel, usually Blue, between -255 and 255</param>
-        public Color(int channel1, int channel2, int channel3)
+        public Color(float channel1, float channel2, float channel3)
         {
             if (channel1 > 255 || channel2 > 255 || channel3 > 255 || channel1 < -255 || channel2 < -255 || channel3 < -255) throw new ArgumentException("Invalid range");
             Channel1 = channel1;
@@ -111,9 +111,9 @@ namespace ImageProcessing.Models
         /// <param name="c">first color</param>
         /// <param name="d">second color</param>
         /// <returns>an array of 3 integers where the elements represent respectively R, G and B</returns>
-        public static int[] operator -(Color c, Color d)
+        public static float[] operator -(Color c, Color d)
         {
-            return new int[] { c.Channel1 - d.Channel1, c.Channel2 - d.Channel2, c.Channel3 - d.Channel3 };
+            return new float[] { c.Channel1 - d.Channel1, c.Channel2 - d.Channel2, c.Channel3 -d.Channel3 };
         }
     }
